add cloud test
